requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
langchain-nvidia-ai-endpoints
langchain-core
aiohttp==3.9.1
beautifulsoup4==4.12.2
PyPDF2==3.0.1
reportlab==4.0.7
lxml==4.9.3
feedparser==6.0.10
langchain-community